move systemtest into specific package
