+++ systemtest/src/main/java/hivemall/systemtest/model/DropTableHQ.java
@@ -1,4 +1,4 @@
-package hivemall.model;
+package hivemall.systemtest.model;
 
 public class DropTableHQ extends TableHQ {
     DropTableHQ(String tableName) {
