+++ systemtest/src/main/java/hivemall/systemtest/model/StrictHQ.java
@@ -1,4 +1,4 @@
-package hivemall.model;
+package hivemall.systemtest.model;
 
 
 abstract public class StrictHQ extends HQ {
