+++ systemtest/src/main/java/hivemall/systemtest/runner/SystemTestRunner.java
@@ -1,17 +1,17 @@
-package hivemall.runner;
+package hivemall.systemtest.runner;
 
-import hivemall.model.CreateTableHQ;
-import hivemall.model.DropTableHQ;
-import hivemall.model.HQ;
-import hivemall.model.InsertHQ;
-import hivemall.model.RawHQ;
-import hivemall.model.StrictHQ;
-import hivemall.model.TableHQ;
-import hivemall.model.TableListHQ;
-import hivemall.model.UploadFileAsNewTableHQ;
-import hivemall.model.UploadFileHQ;
-import hivemall.model.UploadFileToExistingHQ;
-import hivemall.utils.IO;
+import hivemall.systemtest.model.CreateTableHQ;
+import hivemall.systemtest.model.DropTableHQ;
+import hivemall.systemtest.model.HQ;
+import hivemall.systemtest.model.InsertHQ;
+import hivemall.systemtest.model.RawHQ;
+import hivemall.systemtest.model.StrictHQ;
+import hivemall.systemtest.model.TableHQ;
+import hivemall.systemtest.model.TableListHQ;
+import hivemall.systemtest.model.UploadFileAsNewTableHQ;
+import hivemall.systemtest.model.UploadFileHQ;
+import hivemall.systemtest.model.UploadFileToExistingHQ;
+import hivemall.systemtest.utils.IO;
 import org.junit.Assert;
 import org.junit.rules.ExternalResource;
 import org.slf4j.Logger;
